custom-server/requirements.txt
seldon-core==1.6.0
# numpy==1.19.4
joblib==1.1.0
google-cloud-storage==1.41.1
scikit-learn==1.0.2